LPS-105884 Move RolesAdminWebKeys to api module
