+++ modules/apps/roles/roles-admin-api/src/main/java/com/liferay/roles/admin/constants/RolesAdminWebKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.roles.admin.web.internal.constants;
+package com.liferay.roles.admin.constants;
 
 /**
  * @author Drew Brokke
