+++ modules/apps/roles/roles-admin-web/src/main/resources/META-INF/resources/init.jsp
@@ -122,9 +122,9 @@ page import="com.liferay.portlet.usersadmin.search.GroupSearch" %><%@
 page import="com.liferay.portlet.usersadmin.search.OrganizationSearch" %><%@
 page import="com.liferay.product.navigation.personal.menu.BasePersonalMenuEntry" %><%@
 page import="com.liferay.roles.admin.constants.RolesAdminPortletKeys" %><%@
+page import="com.liferay.roles.admin.constants.RolesAdminWebKeys" %><%@
 page import="com.liferay.roles.admin.kernel.util.RolesAdminUtil" %><%@
 page import="com.liferay.roles.admin.role.type.contributor.RoleTypeContributor" %><%@
-page import="com.liferay.roles.admin.web.internal.constants.RolesAdminWebKeys" %><%@
 page import="com.liferay.roles.admin.web.internal.display.context.EditRoleAssignmentsManagementToolbarDisplayContext" %><%@
 page import="com.liferay.roles.admin.web.internal.display.context.RoleDisplayContext" %><%@
 page import="com.liferay.roles.admin.web.internal.display.context.SegmentsEntryDisplayContext" %><%@
